api/requirements.txt
fastapi
beautifulsoup4 
requests
pymilvus
markdown-text-clean
python-dotenv